refactor(shared-core): Improve type safety and session management across services
- Remove unnecessary explicit decimal casts in revenue calculations
- Replace Count property with Count() method for LINQ collections
- Update CurrentUserService to use role-based session timeout instead of inactivity minutes
- Add ICurrentUserService dependency to EnhancedSalesService for user context
- Improve type consistency in DashboardDataAggregationService calculations
- Enhance DashboardService revenue and transaction calculations with proper type handling
- These changes improve code maintainability and ensure consistent type handling across dashboard and sales services

+++ src/Shared.Core/Services/DashboardService.cs
@@ -706,9 +706,9 @@ public class DashboardService : IDashboardService
             var todaySales = await _saleRepository.GetSalesByShopAndDateRangeAsync(shop.Id, analysisDate.StartDate, analysisDate.EndDate);
             var yesterdaySales = await _saleRepository.GetSalesByShopAndDateRangeAsync(shop.Id, analysisDate.StartDate.AddDays(-1), analysisDate.StartDate.AddTicks(-1));
 
-            var todayRevenue = todaySales.Sum(s => (decimal)s.TotalAmount);
-            var yesterdayRevenue = yesterdaySales.Sum(s => (decimal)s.TotalAmount);
-            var todayTransactions = todaySales.Count;
+            var todayRevenue = todaySales.Sum(s => s.TotalAmount);
+            var yesterdayRevenue = yesterdaySales.Sum(s => s.TotalAmount);
+            var todayTransactions = todaySales.Count();
             var averageOrderValue = todayTransactions > 0 ? todayRevenue / (decimal)todayTransactions : 0;
 
             var revenueGrowthPercentage = yesterdayRevenue > 0 
@@ -754,8 +754,8 @@ public class DashboardService : IDashboardService
             var products = await _productRepository.GetProductsByShopAsync(shop.Id);
             var stocks = await _stockRepository.GetStockByShopAsync(shop.Id);
 
-            var revenue = sales.Sum(s => (decimal)s.TotalAmount);
-            var transactionCount = sales.Count;
+            var revenue = sales.Sum(s => s.TotalAmount);
+            var transactionCount = sales.Count();
             var averageOrderValue = transactionCount > 0 ? revenue / (decimal)transactionCount : 0;
             var profitEstimate = revenue * 0.3m; // Estimate
             var inventoryValue = stocks.Sum(s => (decimal)(s.Quantity * (s.Product?.UnitPrice ?? 0)));
